pcds_ioc_builder/requirements.txt
apischema==0.18.0
whatrecord==0.4.2
pyyaml==6.0
